+++ SpikaEnterprise/src/com/clover/spika/enterprise/chat/VideoActivity.java
@@ -189,7 +189,9 @@ public class VideoActivity extends BaseActivity {
 	@Override
 	protected void onDestroy() {
 		super.onDestroy();
-		new File(sFileName).delete();
+		if (sFileName != null) {
+			new File(sFileName).delete();
+		}
 	}
 
 }
