+++ SpikaEnterprise/src/com/clover/spika/enterprise/chat/VoiceActivity.java
@@ -178,7 +178,9 @@ public class VoiceActivity extends BaseActivity {
 	@Override
 	protected void onDestroy() {
 		super.onDestroy();
-		new File(sFileName).delete();
+		if (sFileName != null) {
+			new File(sFileName).delete();
+		}
 	}
 
 }
